+++ src/main/java/com/splitbills/database/models/Debt.java
@@ -7,10 +7,10 @@ public class Debt {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private int id;
+    private Long id;
     private String lender;
     private String debtor;
     private double amount;
-    private int groupId;
+    private Long groupId;
 
 }
