+++ src/main/java/com/splitbills/database/models/PayRecord.java
@@ -10,11 +10,11 @@ public class PayRecord {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private int id;
+    private Long id;
     private String lender;
     private String debtor;
     private double amount;
     private String reason;
-    private int groupId;
+    private Long groupId;
 
 }
